Feature/schedule lifestyle followup (#103)
    Category: implementation
    JIRA issue: MIC-3902
    Research reference: lifestyle intervention ramp

Changes and notes
Initialize last FPG test date and update last FPG test date on timestep cleanup in healthcare utilization component rather than treatment because of component ordering. Some changes to lifestyle ramp in treatment as a result of this.

Verification and Testing
Checked for two timesteps in an interactive context that simulants who were enrolled in lifestyle on that timestep had a scheduled followup.